+++ Code/LYRUIConversationCollectionViewMoreMessagesHeader.h
@@ -1,0 +1,13 @@
+//
+//  LYRUIConversationCollectionViewMoreMessagesHeader.h
+//  LayerUIKit
+//
+//  Created by Ben Blakley on 1/15/15.
+//  Copyright (c) 2015 Layer, Inc. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface LYRUIConversationCollectionViewMoreMessagesHeader : UICollectionReusableView
+
+@end